+++ front/src/Apps/Companies/common/utils.ts
@@ -82,12 +82,6 @@ export const COMPANY_CONSTANTS = [
     label:
       "Installation dans laquelle les déchets perdent leur statut de déchet",
     helpText: ""
-  },
-  {
-    value: CompanyType.Crematorium,
-    label: "Crématorium",
-    helpText:
-      "Un crématorium autorisé prend en charge l'incinération des pièces anatomiques d'origine humaine"
   }
 ];
 
@@ -125,7 +119,7 @@ export const WASTE_PROCESSOR_OPTIONS = [
     value: WasteProcessorType.NonDangerousWastesIncineration
   },
   {
-    label: "Crémation",
+    label: "Crématorium (et cimetières pour la Guyane)",
     value: WasteProcessorType.Cremation
   },
   {